pyproject.toml
[tool.poetry]
name = "explainable-chess-engine"
version = "0.1.0"
description = ""
authors = ["Antoine Déchappe <dechantoine@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.2"
pandas = "^2.1.4"
jupyter = "^1.0.0"
torch = "^2.1.2"
python-chess = "^1.999"
tqdm = "^4.66.1"
torchvision = "^0.17.2"
tensorboard = "^2.16.2"
seaborn = "^0.13.2"
scikit-learn = "^1.4.2"
protobuf = "4.25.3"
pre-commit = "^3.7.0"
pympler = "^1.0.1"
multichoice = "^0.3.5"
more-itertools = "^10.2.0"
anytree = "^2.12.1"
graphviz = "^0.20.3"
gradio = "^4.36.0"
huggingface-hub = "^0.23.3"
stockfish = "^3.28.0"
pyarrow = "^17.0.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
snakeviz = "^2.2.0"

[tool.poetry.group.hf-space.dependencies]
anytree = "^2.12.1"
torch = "^2.1.2"
pandas = "^2.1.4"
python-chess = "^1.999"
graphviz = "^0.20.3"
gradio = "^4.36.0"
more-itertools = "^10.2.0"
loguru = "^0.7.2"
cairosvg = "^2.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
log_cli = true
log_cli_level = 'INFO'
